User menu for mobiles using Drawer
