+++ src/main/ui/platformNavigation/PlatformNavigationBar.tsx
@@ -2,22 +2,30 @@ import NavigationBar from '../../../core/ui/navigation/NavigationBar';
 import PlatformNavigationUserAvatar from './PlatformNavigationUserAvatar';
 import PlatformSearch from '../platformSearch/PlatformSearch';
 import PlatformNavigationMenuButton from './PlatformNavigationMenuButton';
-import { Box } from '@mui/material';
-import { ReactNode } from 'react';
+import { Box, useMediaQuery, Theme } from '@mui/material';
+import React, { ReactNode } from 'react';
+import PlatformNavigationUserMenu from './PlatformNavigationUserMenu';
+import UserMenuPlatformNavigationSegment from './platformNavigationMenu/UserMenuPlatformNavigationSegment';
 
 interface PlatformNavigationBarProps {
   breadcrumbs: ReactNode;
 }
 
 const PlatformNavigationBar = ({ breadcrumbs }: PlatformNavigationBarProps) => {
+  const isMobile = useMediaQuery<Theme>(theme => theme.breakpoints.down('lg'));
+
   return (
     <NavigationBar
       childrenLeft={breadcrumbs}
       childrenRight={
         <Box display="flex" padding={1} gap={1}>
           <PlatformSearch />
-          <PlatformNavigationMenuButton />
-          <PlatformNavigationUserAvatar />
+          {!isMobile && <PlatformNavigationMenuButton />}
+          <PlatformNavigationUserAvatar drawer={isMobile}>
+            <PlatformNavigationUserMenu surface={!isMobile}>
+              {isMobile && <UserMenuPlatformNavigationSegment />}
+            </PlatformNavigationUserMenu>
+          </PlatformNavigationUserAvatar>
         </Box>
       }
     />
