+++ src/main/ui/platformNavigation/PlatformNavigationMenuButton.tsx
@@ -1,16 +1,15 @@
 import React, { Ref } from 'react';
-import { Box, ClickAwayListener, Grow } from '@mui/material';
 import { ReactComponent as AlkemioLogo } from '../logo/logoSmall.svg';
 import NavigationItemContainer from '../../../core/ui/navigation/NavigationItemContainer';
 import { gutters } from '../../../core/ui/grid/utils';
 import NavigationItemButton from '../../../core/ui/navigation/NavigationItemButton';
-import ClickableTooltip from '../../../core/ui/tooltip/ClickableTooltip';
-import PlatformNavigationMenu from './PlatformNavigationMenu';
+import MenuTriggerButton from '../../../core/ui/tooltip/MenuTriggerButton';
+import StandalonePlatformNavigationMenu from './platformNavigationMenu/StandalonePlatformNavigationMenu';
 import { PLATFORM_NAVIGATION_MENU_Z_INDEX } from './constants';
 
 const PlatformNavigationMenuButton = () => {
   return (
-    <ClickableTooltip
+    <MenuTriggerButton
       placement="bottom-end"
       renderTrigger={({ ref, ...props }) => (
         <NavigationItemContainer ref={ref as Ref<HTMLDivElement>}>
@@ -24,21 +23,8 @@ const PlatformNavigationMenuButton = () => {
       )}
       zIndex={PLATFORM_NAVIGATION_MENU_Z_INDEX}
     >
-      {({ onClose, onClickAway, TransitionProps }) => (
-        <Grow
-          {...TransitionProps}
-          style={{
-            transformOrigin: 'right top',
-          }}
-        >
-          <Box padding={gutters(0.5)} paddingRight={0}>
-            <ClickAwayListener onClickAway={onClickAway}>
-              <PlatformNavigationMenu onClose={onClose} />
-            </ClickAwayListener>
-          </Box>
-        </Grow>
-      )}
-    </ClickableTooltip>
+      <StandalonePlatformNavigationMenu />
+    </MenuTriggerButton>
   );
 };
